# internal/api/spec/openapi.yaml
openapi: 3.1.0
info:
  title: pool.biswas.me
  version: "1.0"
  summary: Pool water chemistry, treatment planning and cost-of-ownership tracking.
  description: |
    Every action the web interface performs is one of these calls — there is no
    separate API surface to keep in sync.

    **Authentication.** Send a personal API key as `Authorization: Bearer
    pool_sk_…`, or as `X-API-Key`. Create one under Settings → API keys in the
    app; the plaintext is shown once and cannot be recovered. A key carries a
    `read` or `read,write` scope, enforced on the HTTP method: a read key
    cannot POST, PATCH or DELETE. Browser sessions authenticate the same
    endpoints with a cookie.

    **Chemistry.** Readings are stored as given. Combined chlorine, adjusted
    alkalinity, the Langelier saturation index and the 0–100 water quality
    score are computed server-side from the pool's volume, surface and
    sanitizer, so a test posted through the API is scored identically to one
    typed into the app.

    **For an agent doing the reading.** The photo path is a convenience, not the
    only way in. An agent with a stronger model can do the whole job through
    the ordinary endpoints: `POST /api/tests` with the readings it transcribed,
    `POST /api/attachments` with the photo and that `test_id`, then
    `GET /api/tests/{id}/context` for the brief and `POST
    /api/tests/{id}/insight` with the analysis it wrote. Nothing about that
...
    say what did the reading.

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKeyHeader: []
...
paths:
...
  /api/me/mfa/totp/qr:
    get:
      tags: [Account]
      summary: The enrolment QR
      description: |
        A PNG, served from an authenticated endpoint rather than embedded in
        the page, so the secret does not end up in page source, browser history
        or a screenshot of the URL bar. Gone once enrolment is confirmed.
      responses:
        "200":
          description: The code.
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { description: No enrolment in progress. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer pool_sk_…`"
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
...